Refraction: Merge two services ============================== Merged two servcies PaymentService.php and PaymentsPopulateService.php into one service
