+++ app/Jobs/ProcessPaymentFile.php
@@ -2,7 +2,7 @@
 
 namespace App\Jobs;
 
-use App\Services\PaymentsPopulateService;
+use App\Services\PaymentService;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Queue\Queueable;
 use Illuminate\Queue\InteractsWithQueue;
@@ -27,7 +27,7 @@ class ProcessPaymentFile implements ShouldQueue
     /**
      * Execute the job.
      */
-    public function handle(PaymentsPopulateService $paymentsPopulateService): void
+    public function handle(PaymentService $paymentsPopulateService): void
     {
         $paymentsPopulateService->populatePayments($this->csvData, $this->fileId);
     }
