+++ routes/api.php
@@ -1,7 +1,7 @@
 <?php
 
 use App\Http\Controllers\PaymentController;
-use App\Http\Controllers\PaymentFileController;
+use App\Http\Controllers\FileController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -10,5 +10,5 @@ Route::get('/user', function (Request $request) {
 })->middleware('auth:sanctum');
 
 Route::post('/payments', [PaymentController::class, 'savePaymentRecord']);
-Route::post('/payments/upload', [PaymentFileController::class, 'uploadFile']);
+Route::post('/payments/upload', [FileController::class, 'uploadFile']);
